Update post route to depend on current_user (auth)

+++ routers/posts.py
@@ -7,6 +7,7 @@ from sqlalchemy.orm import selectinload
 import models 
 from asynco.database import get_db
 from schemas import PostResponse, PostCreate, PostUpdate
+from auth import CurrentUser
 
 router = APIRouter()
 
@@ -34,19 +35,12 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
     return post
 
 @router.post("", status_code=status.HTTP_201_CREATED, response_model=PostResponse)
-async def create_post(post: PostCreate, db:Annotated[AsyncSession,Depends(get_db)]):
-    stmt = select(models.User).where(models.User.id == post.user_id)
-    user = (await db.execute(stmt)).scalar_one_or_none ()
-    if not user:
-        raise HTTPException(
-            status_code= status.HTTP_404_NOT_FOUND,
-            detail= "User not found"
-        )
+async def create_post(post: PostCreate, current_user: CurrentUser, db:Annotated[AsyncSession,Depends(get_db)]):
+
     new_post = models.Post(
         title = post.title,
         content = post.content,
-        user_id = post.user_id
-    )
+        user_id = current_user.id)
 
     db.add(new_post)
     await db.commit()
@@ -58,6 +52,7 @@ async def create_post(post: PostCreate, db:Annotated[AsyncSession,Depends(get_db
 async def update_post_full(
     post_id: int, 
     post_data: PostCreate, 
+    current_user: CurrentUser,
     db: Annotated[AsyncSession, Depends(get_db)]
 ):
     post = await db.get(models.Post, post_id)
@@ -65,13 +60,11 @@ async def update_post_full(
     if not post:
         raise HTTPException(status_code=404, detail="Post not Found")
 
-    user = await db.get(models.User, post_data.user_id)
-    if not user:
-        raise HTTPException(status_code=404, detail="User not Found")
-
+    if post.user_id != current_user.id:
+        raise HTTPException (status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
+    
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
 
     await db.commit()
     await db.refresh(post, ["author"])
@@ -82,12 +75,16 @@ async def update_post_full(
 async def update_post_partial(
     post_id: int, 
     post_data: PostUpdate, 
+    current_user: CurrentUser,
     db: Annotated[AsyncSession, Depends(get_db)]
 ):
     post = await db.get(models.Post, post_id)
 
     if not post:
         raise HTTPException(status_code=404, detail="Post not Found")
+    
+    if post.user_id != current_user.id:
+        raise HTTPException (status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this")
 
     update_data = post_data.model_dump(exclude_unset= True) #to remove default values
     for field, value in update_data.items():
@@ -99,11 +96,17 @@ async def update_post_partial(
     return post
 
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db:Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id: int, current_user: CurrentUser, db:Annotated[AsyncSession, Depends(get_db)]):
     post = await db.get(models.Post, post_id)
     if not post: 
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
+        )
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to delete this post"
         )
     await db.delete(post)
     await db.commit()
